+++ module/CLI/src/Command/ShortUrl/CreateShortUrlCommand.php
@@ -134,7 +134,7 @@ class CreateShortUrlCommand extends Command
         }
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output): ?int
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = $this->getIO($input, $output);
         $longUrl = $input->getArgument('longUrl');