+++ module/CLI/src/Command/ShortUrl/DeleteShortUrlCommand.php
@@ -47,7 +47,7 @@ class DeleteShortUrlCommand extends Command
             );
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output): ?int
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
         $identifier = ShortUrlIdentifier::fromCli($input);