+++ module/CLI/src/Command/ShortUrl/ListShortUrlsCommand.php
@@ -129,7 +129,7 @@ class ListShortUrlsCommand extends Command
             );
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output): ?int
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);