+++ module/CLI/src/Command/ShortUrl/ResolveUrlCommand.php
@@ -49,7 +49,7 @@ class ResolveUrlCommand extends Command
         }
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output): ?int
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);