+++ module/CLI/test/Command/Db/CreateDatabaseCommandTest.php
@@ -22,7 +22,7 @@ use ShlinkioTest\Shlink\CLI\Util\CliTestUtils;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\Lock\LockFactory;
-use Symfony\Component\Lock\LockInterface;
+use Symfony\Component\Lock\SharedLockInterface;
 use Symfony\Component\Process\PhpExecutableFinder;
 
 class CreateDatabaseCommandTest extends TestCase
@@ -37,7 +37,7 @@ class CreateDatabaseCommandTest extends TestCase
     protected function setUp(): void
     {
         $locker = $this->createMock(LockFactory::class);
-        $lock = $this->createMock(LockInterface::class);
+        $lock = $this->createMock(SharedLockInterface::class);
         $lock->method('acquire')->withAnyParameters()->willReturn(true);
         $locker->method('createLock')->withAnyParameters()->willReturn($lock);