+++ module/CLI/test/GeoLite/GeolocationDbUpdaterTest.php
@@ -34,7 +34,7 @@ class GeolocationDbUpdaterTest extends TestCase
     {
         $this->dbUpdater = $this->createMock(DbUpdaterInterface::class);
         $this->geoLiteDbReader = $this->createMock(Reader::class);
-        $this->lock = $this->createMock(Lock\LockInterface::class);
+        $this->lock = $this->createMock(Lock\SharedLockInterface::class);
         $this->lock->method('acquire')->with($this->isTrue())->willReturn(true);
     }