+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 * [#1909](https://github.com/shlinkio/shlink/issues/1909) Update docker image to PHP 8.3.
 * [#1786](https://github.com/shlinkio/shlink/issues/1786) Run API tests with RoadRunner by default.
 * [#2008](https://github.com/shlinkio/shlink/issues/2008) Update to Doctrine ORM 3.0.
+* [#2010](https://github.com/shlinkio/shlink/issues/2010) Update to Symfony 7.0 components.
 
 ### Deprecated
 * *Nothing*